markdown adjustment functions are moved out of SFC file
