+++ src/client/components/article.vue
@@ -16,108 +16,7 @@ import VueLayout from "./layout.vue";
 import { defineComponent, nextTick, reactive, ref, watch } from "vue";
 import { useRoute } from "vue-router";
 import hljs from "highlight.js";
-import { fetchPageContent } from "../modules/";
-
-function applyMermaid() {
-  document
-    .querySelectorAll("#article pre code.language-mermaid")
-    .forEach((codeTag) => {
-      const container = codeTag.parentNode;
-      if (container.classList.contains("language-mermaid")) {
-        return;
-      }
-      container.classList.add("language-mermaid");
-      const div = container.appendChild(document.createElement("div"));
-      div.innerHTML = codeTag.innerHTML;
-      div.classList.add("mermaid");
-    });
-  nextTick(() => {
-    // eslint-disable-next-line no-undef
-    mermaid.init();
-  });
-}
-
-function waitMs(ms) {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-function applyCopyable() {
-  document.querySelectorAll("#article pre code.hljs").forEach((codeTag) => {
-    const container = codeTag.parentNode;
-    if (container.classList.contains("copyable")) {
-      return;
-    }
-    container.classList.add("copyable");
-    const button = container.appendChild(document.createElement("button"));
-    button.classList.add("copy-button");
-    const iconTag = button.appendChild(document.createElement("span"));
-    iconTag.classList.add("material-icons", "icon");
-    iconTag.innerHTML = "content_copy";
-    button.addEventListener("click", () => {
-      navigator.clipboard
-        .writeText(codeTag.textContent)
-        .then(() => {
-          iconTag.innerHTML = "done";
-          return waitMs(1000);
-        })
-        .then(() => {
-          iconTag.innerHTML = "content_copy";
-        });
-    });
-  });
-}
-
-function adjustCheckboxes() {
-  document
-    .querySelectorAll("#article li input[type='checkbox']")
-    .forEach((inputTag) => {
-      const listItemTag = inputTag.parentNode;
-      const listTag = listItemTag.parentNode;
-
-      if (listTag.classList.contains("check-list")) {
-        return;
-      }
-      listTag.classList.add("check-list");
-    });
-}
-
-function wrapTable() {
-  document.querySelectorAll("#article table").forEach((tableTag) => {
-    if (!tableTag.parentNode.classList.contains("table-wrapper")) {
-      tableTag.outerHTML = `<div class="table-wrapper scroll-hidden">${tableTag.outerHTML}</div>`;
-    }
-  });
-}
-
-function replaceUrl(originalPath, currentUrl) {
-  if (originalPath.match(/^http/i)) {
-    return originalPath;
-  }
-  const replacedUrl = new URL(originalPath, currentUrl);
-  return replacedUrl.origin + replacedUrl.pathname;
-}
-
-function adjustLinks(currentPage) {
-  document.querySelectorAll("article a").forEach((anchorTag) => {
-    const hyperReference = anchorTag.getAttribute("href");
-    const replacedUrl = replaceUrl(hyperReference, currentPage.url);
-    if (hyperReference === replacedUrl) {
-      return;
-    }
-    anchorTag.setAttribute("href", `#/?path=${replacedUrl}`);
-  });
-}
-
-function adjustImagePaths(currentPage) {
-  document.querySelectorAll("#article img").forEach((imageTag) => {
-    const src = imageTag.getAttribute("src");
-    const replacedSrc = replaceUrl(src, currentPage.url);
-    if (src === replacedSrc) {
-      return;
-    }
-    imageTag.setAttribute("src", replacedSrc);
-  });
-}
+import { fetchPageContent, markdownAdjuster } from "../modules/";
 
 export default defineComponent({
   components: {
@@ -164,12 +63,12 @@ export default defineComponent({
       (to) => {
         nextTick(() => {
           hljs.highlightAll();
-          applyMermaid();
-          applyCopyable();
-          adjustCheckboxes();
-          wrapTable();
-          adjustLinks(state.currentPage);
-          adjustImagePaths(state.currentPage);
+          markdownAdjuster.applyMermaid();
+          markdownAdjuster.applyCopyable();
+          markdownAdjuster.adjustCheckboxes();
+          markdownAdjuster.wrapTable();
+          markdownAdjuster.adjustLinks(state.currentPage);
+          markdownAdjuster.adjustImagePaths(state.currentPage);
           layout?.value?.scrollToTop();
         });
       },
